% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{align_aoa_vec}
\alias{align_aoa_vec}
\title{Reorder age of acquisition (AoA) values to align with network nodes}
\usage{
align_aoa_vec(net, aoa)
}
\arguments{
\item{net}{A binary adjacency matrix.}

\item{aoa}{Named vector of AoA values or a data frame with labels in column 1
and AoA values in column 2.}
}
\value{
A named vector reordered to align with network nodes (column names in
  adjacency matrix).
}
\description{
Reorder age of acquisition (AoA) values to align with network nodes
}
