% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{bayesian_information_criterion}
\alias{bayesian_information_criterion}
\title{Bayesian information criterion (BIC)}
\usage{
bayesian_information_criterion(chisq, df, nobs)
}
\arguments{
\item{chisq}{The chi-square value of the likelihood ratio.}

\item{df}{The degrees of freedom for the likelihood ratio test.}

\item{n}{The number of observations the models were fit to.}
}
\value{
A real scalar
}
\description{
Bayesian information criterion (BIC)
}
