% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{bruteforce_optim}
\alias{bruteforce_optim}
\title{Brute force optimization}
\usage{
bruteforce_optim(formula, data, split_by, R = 10000)
}
\arguments{
\item{formula}{A formula; see details}

\item{data}{A data frame containing, at least, the variables implied by the
formula.}

\item{split_by}{A string referencing a variable in \code{data} that encodes
time steps; see details.}

\item{R}{The number of random models to evaluate.}
}
\value{
A vector expressing the best model parameters discovered in the brute
  force search.
}
\description{
Evaluate a large number of randomly generated models. Useful for obtaining an
initial model to begin a gradient optimization.
}
\details{
The formula must reference a variable that encodes the time-point at which
each node is acquired as the dependent variable. If \code{data}
contains the data for multiple time steps, then this value should probably be
the same for all points in time at which the node is referenced (although the
function does not enforce this). The independent variables specify the set of
variables that will be used to predict, at each point in time, whether the
node is acquired or not.

The \code{split_by} variable is used to split \code{data} by time-step, so
that probabilities can be estimated independently at discrete points in time.
The assumption is that a node will be acquired once, at a single point in
time.
}
