% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{gradient_optim}
\alias{gradient_optim}
\title{Gradient optimization of the negative log likelihood}
\usage{
gradient_optim(formula, data, B0, split_by, control = list(maxit = 1000))
}
\arguments{
\item{formula}{A formula; see details}

\item{data}{A data frame containing, at least, the variables implied by the
formula.}

\item{B0}{Initial values for the model parameters.}

\item{split_by}{A string referencing a variable in \code{data} that encodes
time steps; see details.}

\item{control}{See documentation for \code{\link[stats]{optim}}.}
}
\value{
See documentation for \code{\link[stats]{optim}}.
}
\description{
Nelder-Mead optimization using the \code{\link[stats]{optim}} function.
}
\details{
The optimization minimizes the negative log likelihood associated with the
estimated probabilities for each nodes at the time it is added to the
network.

The formula must reference a variable that encodes the time-point at which
each node is acquired as the dependent variable. If \code{data}
contains the data for multiple time steps, then this value should probably be
the same for all points in time at which the node is referenced (although the
function does not enforce this). The independent variables specify the set of
variables that will be used to predict, at each point in time, whether the
node is acquired or not.

The \code{split_by} variable is used to split \code{data} by time-step, so
that probabilities can be estimated independently at discrete points in time.
The assumption is that a node will be acquired once, at a single point in
time.
}
