% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{growth_values}
\alias{growth_values}
\title{Compute growth values based on network structure}
\usage{
growth_values(
  net,
  aoa_tbl,
  growth_models = "all",
  weighted = FALSE,
  alpha = 0.5
)
}
\arguments{
\item{net}{A binary adjacency matrix.}

\item{aoa_tbl}{A vector of known vertexes. Can be a logical vector.}

\item{weighted}{Logical. Should weights be taken into account?}

\item{alpha}{What proportion of the growth value should be influenced by the
weight [1]? This is ignored if \code{weighted} is FALSE.}
}
\value{
A vector of growth values for unknown words.

Growth values are computed for the words represented by each column. If the
model is not expressed by a symmetric matrix, the column names are used to
filter the rows so that it is made to be symmetric when computing growth values
according to the lure of the associates and preferential attechment. These
models are concerned with only the words that can be learned---and in the
context of these growth values, only words that have AoA values can be learned.
So the columns correspond to words for which we have AoA values, and that
defines the scope of the growth model.

The preferential acquisition growth model, unlike the other two, takes into
account the structure of the whole environment, not limited to only those words
that can be learned. In other words, the indegree of a word that can be learned
can be influenced by words that are in the environment but cannot be learned.
Thus, the model can be asymmetric, such that it has more rows than columns,
and this will change the pattern of growth predicted by preferential
acquisition.
}
\description{
Compute growth values based on network structure
}
\references{
1. Opsahl, T., Agneessens, F., & Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245-251. doi:https://doi.org/10.1016/j.socnet.2010.03.006
}
\seealso{
\code{\link[netgrowr]{PreferentialAttachment}}, \code{\link[netgrowr]{LureOfTheAssociates}}, \code{\link[netgrowr]{PreferentialAcquisition}}, \code{\link[netgrowr]{opsahl_weighted_degree}}
}
