% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{is_known}
\alias{is_known}
\title{Flag words acquired before a given month}
\usage{
is_known(month, aoa)
}
\arguments{
\item{month}{Positive integer indicating a reference month.}

\item{aoa}{Vector of age of acquisition values}
}
\value{
Logical vector

A word flagged as "known" entered the vocabulary prior to the indicated
month.
}
\description{
Flag words acquired before a given month
}
