% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{is_learned}
\alias{is_learned}
\title{Flag words acquired during a given month}
\usage{
is_learned(month, aoa)
}
\arguments{
\item{month}{Positive integer indicating a reference month.}

\item{aoa}{Vector of age of acquisition values}
}
\value{
Logical vector

A word flagged as "learned" enters the vocabulary in the indicated month.
}
\description{
Flag words acquired during a given month
}
