% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log likelihood}
\usage{
loglikelihood(p)
}
\arguments{
\item{p}{A vector or matrix of probabilities; see details.}
}
\value{
If \code{p} is a matrix, a vector of log likelihoods as long as \code{ncol(p)}, otherwise a single value.
}
\description{
Log likelihood
}
\details{
If \code{p} is a matrix, each column represents the probabilities associated with a different model.
}
