% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{model_comparison}
\alias{model_comparison}
\title{Compare two models with likelihood ratio test}
\usage{
model_comparison(full, restricted)
}
\arguments{
\item{full}{list object associated with the full model.}

\item{restricted}{list object associated with the restricted (i.e., nested) model.}
}
\value{
A named vector with the following values:
\item{p}{One-tailed probability that the likelihood ratio is sampled from a
chi-square distribution with \code{df} degrees of freedom centered on zero or
less.}
\item{df}{The degrees of freedom of the statistical test; the difference in
the number of parameters between the nested models.}
\item{L0}{The negative log-likelihood of the restricted model.}
\item{L1}{The negative log-likelihood of the full model.}
\item{chisq}{The chi-square value of the likelihood ratio.}
\item{BIC}{The Bayesian Information Criterion for the model comparison, based on the chi-squared likelihood ratio.}
}
\description{
Compare two models with likelihood ratio test
}
