% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{model_frame_with_split_and_labels}
\alias{model_frame_with_split_and_labels}
\title{Return model frame with split and and label variables}
\usage{
model_frame_with_split_and_labels(formula, split_by, label_with, data)
}
\arguments{
\item{formula}{a formula}

\item{split_by}{A string referencing a variable in \code{data} that encodes
time steps; see details for \code{\link[netgrowr]{mle_network_growth}}.}

\item{label_with}{A string referencing a variable in \code{data} that
provides node-labels. Labels may repeat across time steps specified by
\code{split_by}.}

\item{data}{A data frame containing, at least, the variables implied by the
formula.}
}
\value{
The model frame
}
\description{
Written to handle the case where either split or labels are not provided.
}
