% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{nterms}
\alias{nterms}
\title{Count terms (independent variables) implied by a function}
\usage{
nterms(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
The number of terms the formula implies.
}
\description{
Unless the formula explicitly states that an intercept term is not desired,
the returned count includes an intercept term.
}
