% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{opsahl_weighted_degree}
\alias{opsahl_weighted_degree}
\title{Re-weight degree as combination of weighted and unweighted values}
\usage{
opsahl_weighted_degree(u, w, alpha)
}
\arguments{
\item{u}{Degree as derived from network with weights discarded (unweighted)}

\item{w}{Degree as derived from network with weights included (weighted)}

\item{alpha}{Mixing parameter, where zero means to only consider \code{u} and one means to only consider \code{w}. Use 0.5 to consult both sources equally.}
}
\value{
Weighted degree
}
\description{
Re-weight degree as combination of weighted and unweighted values
}
\references{
1. Opsahl, T., Agneessens, F., & Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245-251. doi:https://doi.org/10.1016/j.socnet.2010.03.006
}
