% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{preferential_acquisition}
\alias{preferential_acquisition}
\title{Compute preferential acquisition values}
\usage{
preferential_acquisition(net, known, weighted = FALSE, alpha = 0.5)
}
\arguments{
\item{net}{A binary adjacency matrix.}

\item{known}{A vector of known vertexes. Can be a logical vector.}

\item{weighted}{Logical. Should weights be taken into account?}

\item{alpha}{What proportion of the growth value should be influenced by the
weight [1]? This is ignored if \code{weighted} is FALSE.}
}
\value{
A vector of growth values for unknown words.

Words that are flagged as \code{known} in the input will be associated with
\code{NA} in the output.

Children acquire words that are hubs in their learning environment,
regardless of what they currently know [2].

Growth value is the degree of the word in the presumed learning environment.
}
\description{
Compute preferential acquisition values
}
\references{
1. Opsahl, T., Agneessens, F., & Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245-251. doi:https://doi.org/10.1016/j.socnet.2010.03.006

2. Hills, T. T., Maouene, M., Maouene, J., Sheya, A., & Smith, L. (2009). Longitudinal analysis of early semantic networks: Preferential attachment or preferential acquisition? Psychological science, 20(6), 729-739. doi:10.1111/j.1467-9280.2009.02365.x
}
\seealso{
\code{\link[netgrowr]{GrowthValue}}, \code{\link[netgrowr]{PreferentialAttachment}}, \code{\link[netgrowr]{LureOfTheAssociates}}, \code{\link[netgrowr]{opsahl_weighted_degree}}
}
