% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-values.R
\name{preferential_attachment}
\alias{preferential_attachment}
\title{Compute preferential attachment values}
\usage{
preferential_attachment(net, known, weighted = FALSE, alpha = 0.5)
}
\arguments{
\item{net}{A binary adjacency matrix.}

\item{known}{A vector of known vertexes. Can be a logical vector.}

\item{weighted}{Logical. Should weights be taken into account?}

\item{alpha}{What proportion of the growth value should be influenced by the
weight [1]? This is ignored if \code{weighted} is FALSE.}
}
\value{
A vector of growth values for unknown words.

Words that are flagged as \code{known} in the input will be associated with
\code{NA} in the output.

Children acquire words that are related to high-degree words that they
already know.

Growth value is the mean degree of the known words that the new word
attached to [2].
}
\description{
Compute preferential attachment values
}
\references{
1. Opsahl, T., Agneessens, F., & Skvoretz, J. (2010). Node centrality in weighted networks: Generalizing degree and shortest paths. Social Networks, 32(3), 245-251. doi:https://doi.org/10.1016/j.socnet.2010.03.006

2. Steyvers, M., & Tenenbaum, J. B. (2005). The large‐scale structure of semantic networks: Statistical analyses and a model of semantic growth. Cognitive science, 29(1), 41-78.
}
\seealso{
\code{\link[netgrowr]{GrowthValue}}, \code{\link[netgrowr]{LureOfTheAssociates}}, \code{\link[netgrowr]{PreferentialAcquisition}}, \code{\link[netgrowr]{opsahl_weighted_degree}}, \code{\link[netgrowr]{opsahl_weighted_degree}}
}
