% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{probability_node_added}
\alias{probability_node_added}
\title{Generate predicted acquisition probabilities for network nodes}
\usage{
probability_node_added(beta, formula, data, split_by, label_with = NULL)
}
\arguments{
\item{beta}{A vector of weights (i.e, the model), or a matrix where each
column is a different model.}

\item{formula}{A formula; see details}

\item{data}{A data frame containing, at least, the variables implied by the
formula.}

\item{split_by}{A string referencing a variable in \code{data} that encodes
time steps; see details.}

\item{label_with}{A string referencing a variable in \code{data} that
provides node-labels. Labels may repeat across time steps specified by
\code{split_by}.}
}
\value{
A numeric vector or matrix; see details.
}
\description{
Compute the ratio of strengths to evaluate the model evidence for acquiring a
node at the time it was acquired. The \code{split_by} parameter
specifies a variable in \code{data} that specifies different points in time.
}
\details{
The formula must reference a variable that encodes the time-point at which
each node is acquired as the dependent variable. If \code{data}
contains the data for multiple time steps, then this value should probably be
the same for all points in time at which the node is referenced (although the
function does not enforce this). The independent variables specify the set of
variables that will be used to predict, at each point in time, whether the
node is acquired or not.

The \code{split_by} variable is used to split \code{data} by time-step, so
that probabilities can be estimated independently at discrete points in time.
The assumption is that a node will be acquired once, at a single point in
time.

The function will return a vector if a beta was a vector or one-column
matrix. Otherwise, the return value will be a matrix, with a row for each
node and a column for each model specified by \code{beta}. Regardless, the
returned values are probabilities for nodes that are acquired, as predicted
at the point in time the node was acquired. Probabilities are computed as a
ratio of strengths.
}
\seealso{
\code{\link[netgrowr]{ratio_of_strengths}}
}
