% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-network-growth.R
\name{ratio_of_strengths}
\alias{ratio_of_strengths}
\title{Compute ratio of strengths}
\usage{
ratio_of_strengths(data, formula, beta)
}
\arguments{
\item{data}{A data frame containing, at least, the variables implied by the
formula.}

\item{formula}{A formula; see details}

\item{beta}{A vector of weights (i.e, the model), or a matrix where each
column is a different model.}
}
\value{
A matrix of probabilities associated with nodes that were
  acquired. There will be a column for each model specified by \code{beta}.
}
\description{
Standardize the evidence for the targets by dividing by the sum of the
evidence across targets and distractors.
}
\details{
The formula must reference a variable that encodes the time-point at which
each node is acquired as the dependent variable. If \code{data}
contains the data for multiple time steps, then this value should probably be
the same for all points in time at which the node is referenced (although the
function does not enforce this). The independent variables specify the set of
variables that will be used to predict, at each point in time, whether the
node is acquired or not.
}
